MyProject.Test/API.Test/Controller.Test/PermissionControllerTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Mvc;
using Moq;
using MyProject.Api.Controllers;
using MyProject.Application.Interfaces;
using MyProject.Domain.Entities;
using Xunit;

namespace MyProject.Tests.Controllers
{
    public class PermissionControllerTests
    {
        [Fact]
        public async Task GetAllPermissions_ReturnsOkResult()
        {
            // Arrange
            var mockService = new Mock<IPermissionService>();
            mockService.Setup(service => service.GetAllAsync()).ReturnsAsync(new List<Permission>());

            var controller = new PermissionController(mockService.Object);

            // Act
            var result = await controller.GetAllPermissions();

            // Assert
            Assert.IsType<OkObjectResult>(result);
        }

        [Theory]
        [InlineData(1)]
        [InlineData(2)]
        public async Task GetPermissionById_ExistingId_ReturnsOkResult(int permissionId)
        {
            // Arrange
            var mockService = new Mock<IPermissionService>();
            mockService.Setup(service => service.GetByIdAsync(permissionId)).ReturnsAsync(new Permission { Id = permissionId, Name = "TestPermission" });

            var controller = new PermissionController(mockService.Object);

            // Act
            var result = await controller.GetPermissionById(permissionId);

            // Assert
            Assert.IsType<OkObjectResult>(result);
        }

        [Fact]
        public async Task CreatePermission_ValidObject_ReturnsCreatedAtAction()
        {
            // Arrange
            var mockService = new Mock<IPermissionService>();
            var newPermission = new Permission { Name = "NewPermission" }; // Replace with test data
            var createdPermissionId = 1; // Replace with an expected ID for testing purposes
            mockService.Setup(service => service.CreateAsync(newPermission)).ReturnsAsync(createdPermissionId);

            var controller = new PermissionController(mockService.Object);

            // Act
            var result = await controller.CreatePermission(newPermission) as CreatedAtActionResult;

            // Assert
            Assert.NotNull(result);
            Assert.Equal(nameof(PermissionController.GetPermissionById), result.ActionName);
            Assert.Equal(createdPermissionId, result.RouteValues["id"]);
        }

        // Write similar tests for UpdatePermission and DeletePermission actions

        // Remember to add specific assertions to validate the responses and behaviors of each action.
    }
}
